producer/requirements.txt
kafka-python
feedparser
pymongo
